requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==5.1.5
django-cors-headers==4.6.0
django-filter==24.3
django-use-email-as-username==1.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-excel==2.5.2
drf-renderer-xlsx==1.0.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
et_xmlfile==2.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
openpyxl==3.1.5
pillow==11.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
XlsxWriter==3.2.0
xlwt==1.3.0
